Added: Nachrichten werden in der DB gespeichert. 2 Methoden eine für die die man versendet, eine weitere für die die man empfängt

+++ MMFT/DB/Models/Nachrichten.cs
@@ -9,7 +9,7 @@ public partial class Nachrichten
 
     public string SUuid { get; set; }
 
-    public int Zeitstempel { get; set; }
+    public long Zeitstempel { get; set; }
 
     public string TInhalt { get; set; }
